requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asn1crypto==1.5.1
asttokens==2.4.1
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
cryptography==42.0.7
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
orjson==3.10.3
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyOpenSSL==24.1.0
pypdf==4.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq==26.0.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
spyder-kernels==2.5.1
stack-data==0.6.3
starlette==0.37.2
tornado==6.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wurlitzer==3.1.0
zipp==3.18.2
